--- backend/ht_buses_app/data/students/students.csv
@@ -1,203 +1,0 @@
-﻿first_name,last_name,student_school_id,school_id,route_id,user_id
-Student 5,Tester 5,123,5,1,8
-Student 6,Tester 6,123,5,1,8
-Student 7,Tester 7,123,5,1,8
-Student 8,Tester 8,123,5,1,8
-Student 9,Tester 9,123,5,1,8
-Student 10,Tester 10,123,5,1,8
-Student 11,Tester 11,123,5,1,8
-Student 12,Tester 12,123,5,1,8
-Student 13,Tester 13,123,5,1,8
-Student 14,Tester 14,123,5,1,8
-Student 15,Tester 15,123,5,1,8
-Student 16,Tester 16,123,5,1,8
-Student 17,Tester 17,123,5,1,8
-Student 18,Tester 18,123,5,1,8
-Student 19,Tester 19,123,5,1,8
-Student 20,Tester 20,123,5,1,8
-Student 21,Tester 21,123,5,1,8
-Student 22,Tester 22,123,5,1,8
-Student 23,Tester 23,123,5,1,8
-Student 24,Tester 24,123,5,1,8
-Student 25,Tester 25,123,5,1,8
-Student 26,Tester 26,123,5,1,8
-Student 27,Tester 27,123,5,1,8
-Student 28,Tester 28,123,5,1,8
-Student 29,Tester 29,123,5,1,8
-Student 30,Tester 30,123,5,1,8
-Student 31,Tester 31,123,5,1,8
-Student 32,Tester 32,123,5,1,8
-Student 33,Tester 33,123,5,1,8
-Student 34,Tester 34,123,5,1,8
-Student 35,Tester 35,123,5,1,8
-Student 36,Tester 36,123,5,1,8
-Student 37,Tester 37,123,5,1,8
-Student 38,Tester 38,123,5,1,8
-Student 39,Tester 39,123,5,1,8
-Student 40,Tester 40,123,5,1,8
-Student 41,Tester 41,123,5,1,8
-Student 42,Tester 42,123,5,1,8
-Student 43,Tester 43,123,5,1,8
-Student 44,Tester 44,123,5,1,8
-Student 45,Tester 45,123,5,1,8
-Student 46,Tester 46,123,5,1,8
-Student 47,Tester 47,123,5,1,8
-Student 48,Tester 48,123,5,1,8
-Student 49,Tester 49,123,5,1,8
-Student 50,Tester 50,123,5,1,8
-Student 51,Tester 51,123,5,1,8
-Student 52,Tester 52,123,5,1,8
-Student 53,Tester 53,123,5,1,8
-Student 54,Tester 54,123,5,1,8
-Student 55,Tester 55,123,5,1,8
-Student 56,Tester 56,123,5,1,8
-Student 57,Tester 57,123,5,1,8
-Student 58,Tester 58,123,5,1,8
-Student 59,Tester 59,123,5,1,8
-Student 60,Tester 60,123,5,1,8
-Student 61,Tester 61,123,5,1,8
-Student 62,Tester 62,123,5,1,8
-Student 63,Tester 63,123,5,1,8
-Student 64,Tester 64,123,5,1,8
-Student 65,Tester 65,123,5,1,8
-Student 66,Tester 66,123,5,1,8
-Student 67,Tester 67,123,5,1,8
-Student 68,Tester 68,123,5,1,8
-Student 69,Tester 69,123,5,1,8
-Student 70,Tester 70,123,5,1,8
-Student 71,Tester 71,123,5,1,8
-Student 72,Tester 72,123,5,1,8
-Student 73,Tester 73,123,5,1,8
-Student 74,Tester 74,123,5,1,8
-Student 75,Tester 75,123,5,1,8
-Student 76,Tester 76,123,5,1,8
-Student 77,Tester 77,123,5,1,8
-Student 78,Tester 78,123,5,1,8
-Student 79,Tester 79,123,5,1,8
-Student 80,Tester 80,123,5,1,8
-Student 81,Tester 81,123,5,1,8
-Student 82,Tester 82,123,5,1,8
-Student 83,Tester 83,123,5,1,8
-Student 84,Tester 84,123,5,1,8
-Student 85,Tester 85,123,5,1,8
-Student 86,Tester 86,123,5,1,8
-Student 87,Tester 87,123,5,1,8
-Student 88,Tester 88,123,5,1,8
-Student 89,Tester 89,123,5,1,8
-Student 90,Tester 90,123,5,1,8
-Student 91,Tester 91,123,5,1,8
-Student 92,Tester 92,123,5,1,8
-Student 93,Tester 93,123,5,1,8
-Student 94,Tester 94,123,5,1,8
-Student 95,Tester 95,123,5,1,8
-Student 96,Tester 96,123,5,1,8
-Student 97,Tester 97,123,5,1,8
-Student 98,Tester 98,123,5,1,8
-Student 99,Tester 99,123,5,1,8
-Student 100,Tester 100,123,5,1,8
-Student 101,Tester 101,123,5,1,8
-Student 102,Tester 102,123,5,1,8
-Student 103,Tester 103,123,5,1,8
-Student 104,Tester 104,123,5,1,8
-Student 105,Tester 105,123,5,1,8
-Student 106,Tester 106,123,5,1,8
-Student 107,Tester 107,123,5,1,8
-Student 108,Tester 108,123,5,1,8
-Student 109,Tester 109,123,5,1,8
-Student 110,Tester 110,123,5,1,8
-Student 111,Tester 111,123,5,1,8
-Student 112,Tester 112,123,5,1,8
-Student 113,Tester 113,123,5,1,8
-Student 114,Tester 114,123,5,1,8
-Student 115,Tester 115,123,5,1,8
-Student 116,Tester 116,123,5,1,8
-Student 117,Tester 117,123,5,1,8
-Student 118,Tester 118,123,5,1,8
-Student 119,Tester 119,123,5,1,8
-Student 120,Tester 120,123,5,1,8
-Student 121,Tester 121,123,5,1,8
-Student 122,Tester 122,123,5,1,8
-Student 123,Tester 123,123,5,1,8
-Student 124,Tester 124,123,5,1,8
-Student 125,Tester 125,123,5,1,8
-Student 126,Tester 126,123,5,1,8
-Student 127,Tester 127,123,5,1,8
-Student 128,Tester 128,123,5,1,8
-Student 129,Tester 129,123,5,1,8
-Student 130,Tester 130,123,5,1,8
-Student 131,Tester 131,123,5,1,8
-Student 132,Tester 132,123,5,1,8
-Student 133,Tester 133,123,5,1,8
-Student 134,Tester 134,123,5,1,8
-Student 135,Tester 135,123,5,1,8
-Student 136,Tester 136,123,5,1,8
-Student 137,Tester 137,123,5,1,8
-Student 138,Tester 138,123,5,1,8
-Student 139,Tester 139,123,5,1,8
-Student 140,Tester 140,123,5,1,8
-Student 141,Tester 141,123,5,1,8
-Student 142,Tester 142,123,5,1,8
-Student 143,Tester 143,123,5,1,8
-Student 144,Tester 144,123,5,1,8
-Student 145,Tester 145,123,5,1,8
-Student 146,Tester 146,123,5,1,8
-Student 147,Tester 147,123,5,1,8
-Student 148,Tester 148,123,5,1,8
-Student 149,Tester 149,123,5,1,8
-Student 150,Tester 150,123,5,1,8
-Student 151,Tester 151,123,5,1,8
-Student 152,Tester 152,123,5,1,8
-Student 153,Tester 153,123,5,1,8
-Student 154,Tester 154,123,5,1,8
-Student 155,Tester 155,123,5,1,8
-Student 156,Tester 156,123,5,1,8
-Student 157,Tester 157,123,5,1,8
-Student 158,Tester 158,123,5,1,8
-Student 159,Tester 159,123,5,1,8
-Student 160,Tester 160,123,5,1,8
-Student 161,Tester 161,123,5,1,8
-Student 162,Tester 162,123,5,1,8
-Student 163,Tester 163,123,5,1,8
-Student 164,Tester 164,123,5,1,8
-Student 165,Tester 165,123,5,1,8
-Student 166,Tester 166,123,5,1,8
-Student 167,Tester 167,123,5,1,8
-Student 168,Tester 168,123,5,1,8
-Student 169,Tester 169,123,5,1,8
-Student 170,Tester 170,123,5,1,8
-Student 171,Tester 171,123,5,1,8
-Student 172,Tester 172,123,5,1,8
-Student 173,Tester 173,123,5,1,8
-Student 174,Tester 174,123,5,1,8
-Student 175,Tester 175,123,5,1,8
-Student 176,Tester 176,123,5,1,8
-Student 177,Tester 177,123,5,1,8
-Student 178,Tester 178,123,5,1,8
-Student 179,Tester 179,123,5,1,8
-Student 180,Tester 180,123,5,1,8
-Student 181,Tester 181,123,5,1,8
-Student 182,Tester 182,123,5,1,8
-Student 183,Tester 183,123,5,1,8
-Student 184,Tester 184,123,5,1,8
-Student 185,Tester 185,123,5,1,8
-Student 186,Tester 186,123,5,1,8
-Student 187,Tester 187,123,5,1,8
-Student 188,Tester 188,123,5,1,8
-Student 189,Tester 189,123,5,1,8
-Student 190,Tester 190,123,5,1,8
-Student 191,Tester 191,123,5,1,8
-Student 192,Tester 192,123,5,1,8
-Student 193,Tester 193,123,5,1,8
-Student 194,Tester 194,123,5,1,8
-Student 195,Tester 195,123,5,1,8
-Student 196,Tester 196,123,5,1,8
-Student 197,Tester 197,123,5,1,8
-Student 198,Tester 198,123,5,1,8
-Student 199,Tester 199,123,5,1,8
-Student 200,Tester 200,123,5,1,8
-Student 201,Tester 201,123,5,1,8
-Student 202,Tester 202,123,5,1,8
-Student 203,Tester 203,123,5,1,8
-Student 204,Tester 204,123,5,1,8
-Student 205,Tester 205,123,5,1,8
-Student 206,Tester 206,123,5,1,8
