+++ backend/ht_buses_app/data/tester_school.csv
@@ -1,0 +1,2 @@
+﻿name,address,arrivalTime,departureTime
+Lyons Preparatory,"222 Holiday Circle, Savannah, GA 31419",7:45,15:30
