+++ backend/ht_buses_app/google_funcs.py
@@ -2,8 +2,8 @@ import googlemaps
 import os
 
 gmaps = googlemaps.Client(key='AIzaSyDKBCCbU8w69KvLefPmHtFRTsP6KAFXp0s')
-def geocode_address(addresses):
+def geocode_address(address):
     geocodes = []
-    for address in addresses:
-        geocodes.append(gmaps.geocode(address)[0]['geometry']['location'])
+    #for address in addresses:
+    geocodes.append(gmaps.geocode(address)[0]['geometry']['location'])
     return geocodes
